+++ MAPH/controller/UserContext.js
@@ -1,0 +1,17 @@
+import React, { createContext, useState } from 'react';
+
+// Crea el contexto
+export const UserContext = createContext();
+
+// Proveedor del contexto
+export const UserProvider = ({ children }) => {
+  const [user, setUser] = useState(null);
+
+  return (
+    <UserContext.Provider value={{ user, setUser }}>
+      {children}
+    </UserContext.Provider>
+  );
+};
+
+// const response = await axios.get(http://localhost:3001/user/${cedula});
