+++ MAPH/package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@expo/metro-runtime": "~4.0.0",
     "@expo/vector-icons": "^14.0.4",
+    "@react-native-mask-input": "^X.X.X",
     "@react-native-community/datetimepicker": "^8.2.0",
     "@react-native-community/viewpager": "^5.0.11",
     "@react-native-picker/picker": "^2.11.0",
@@ -18,21 +19,33 @@
     "@react-navigation/native": "^7.0.14",
     "@react-navigation/stack": "^7.1.1",
     "axios": "^1.7.9",
+    "browser-image-compression": "^2.0.2",
     "cors": "^2.8.5",
     "expo": "~52.0.25",
+    "expo-document-picker": "~13.0.2",
+    "expo-image-picker": "~16.0.4",
     "expo-status-bar": "~2.0.1",
+    "express": "^4.21.2",
+    "multer": "^1.4.5-lts.1",
+    "mysql2": "^3.12.0",
     "react": "18.3.1",
     "react-dom": "18.3.1",
+    "react-hook-form": "^7.54.2",
     "react-native": "0.76.6",
     "react-native-calendars": "^1.1308.0",
+    "react-native-document-picker": "^9.3.1",
     "react-native-gesture-handler": "^2.22.0",
+    "react-native-image-picker": "^7.2.3",
+    "react-native-mask-input": "^1.2.3",
     "react-native-modal-datetime-picker": "^18.0.0",
     "react-native-reanimated": "^3.16.7",
     "react-native-safe-area-context": "^5.1.0",
     "react-native-screens": "^4.5.0",
+    "react-native-text-input-mask": "^3.2.0",
     "react-native-toast-message": "^2.2.1",
     "react-native-vector-icons": "^10.2.0",
-    "react-native-web": "~0.19.13"
+    "react-native-web": "~0.19.13",
+    "rn-fetch-blob": "^0.12.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
